leveldb: move leveldb into its own package
