+++ leveldb/leveldb_suite_test.go
@@ -1,0 +1,1 @@
+package leveldb_test
