+++ leveldb/leveldb_test.go
@@ -1,4 +1,4 @@
-package store_test
+package leveldb_test
 
 import (
 	"bytes"
